edit tests files
